Izmijenjene python aplikacije, dovršeno učitavanje datoteka
Python kodovi su pretvoreni u funkcije da se mogu pozvati iz app.py, datoteke su također premještene da budu u istoj mapi kao i app.py.
Dovršena implementacija učitavanja i obrade datoteka - datoteke se prilikom klika na gumb stavljaju u mapu 'uploads', podaci se prenose u bazu podataka, te se na kraju datoteke brišu iz 'uploads' mape.
